+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 ### Added
 - Add a default `.editorconfig` to help enforce coding conventions
+- Added Partner app generator via a new PartnerAppService and rake task
 
 ## 0.13.6 - 2017-06-21
 ### Added
@@ -16,7 +17,6 @@ All notable changes to this project will be documented in this file.
 - Added Rollbar "Person tracking" for `ShopJob`
 - Added overrides to the `data` attribute for `Synchronises` models to allow
   accessing the `data` hash with indifferent access.
-- Added Partner app generator via a new PartnerAppService and rake task
 
 ### Changed
 - Turn prepared statements off by default in config/database.yml template